SAMSON-158 samsonModuleParser detects if there has been a change in key and value formats between module and implementation. It also uses now comments instead of ifdef when keeping this information in .h file. Many modules have been corrected
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1922 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ modules/cdr/operations/convert_cdrs_gst2mob.h
@@ -25,7 +25,7 @@ namespace cdr{
 // If interface changes and you do not recreate this file, consider updating this information (and of course, the module file)
 
 input: system.UInt cdr.gstCdr  
-output: system.UInt64 cdr.mobCdr
+output: system.UInt cdr.mobCdr
 
 helpLine: Convert CDRs from Gasset format to Mob format.
 extendedHelp: 		Convert CDRs from Gasset format to Mob format.
